+++ css/greenberg.css
@@ -1,0 +1,4 @@
+#storyArea {
+    max-height: 400px;
+    overflow-y: auto;
+}
