+++ css/index.css
@@ -1,0 +1,10 @@
+body {
+    /*background-image: url("../img/heartwelcome.jpg");*/
+    background-repeat: no-repeat;
+    background-attachment: fixed;
+    background-position-x: center;
+    background-position-y: top;
+    background-size: 100%;
+    margin-bottom: 0px;
+    margin-top: 0px;
+}
