+++ index.php
@@ -1,37 +1,24 @@
-<?php
-     include("mm2.php");
-?>
 <html>
-  
   <head>
     <?php require("includes/head.php") ?>
+    <meta charset="utf-8" />
+    <title>Izzy's Epicality</title>
+    <link rel="stylesheet" type="text/css" href="css/index.css" />
  </head>
-
-  <title>Izzy's Epicailty</title>
-  <style>
-    body  {
-        background-image: url("img/heartwelcome.jpg");
-        background-repeat: no-repeat;
-        background-attachment: fixed;
-        background-position-x: center;
-        background-position-y: top;   
-        background-size: 100%;
-        margin-bottom: 0px;
-        margin-top: 0px;
-         }
-         
-  </style>
  <body class="bloodyeverything">
-    
-      <footer>
-      <div class="footer col-sm-6">
-        <footer>A website dedicated to Izzy's deep dark holes of inspiration.</footer>
+  <div class="container-fluid">
+      <?php require("navigation.php") ?>
+    <div class="row">
+      <div class="footer col-md-6">
+          A website dedicated to Izzy's deep dark holes of inspiration.
       </div>
-      <div class="footer2 col-sm-6">
-        <footer>LAST POST: Greenberg's Experiment at January 2nd, 2017</footer>
+      <div class="footer2 col-md-6">
+          LAST POST: Greenberg's Experiment at January 2nd, 2017
       </div> 
-      </footer>
+    </div>
+  </div>
   </body>
+</html>
